Support for Gatling 3.0.0
In Gatling 3.0.0, the REQUEST line no longer has the scenario. To
determine the scenario, you need to get the userId, then lookup
the scenario associated with the user.

Fix is a little ugly, relying on `getUserType` being invoked, but
otherwise it would require more extensive changes.

Some of the other indexes were also wrong, updated them.